+++ example/lib/main.dart
@@ -1,7 +1,9 @@
 import 'package:impaktfull_ui/impaktfull_ui.dart';
 import 'package:impaktfull_ui_example/src/screen/home_screen.dart';
 
-void main() => runApp(const ImpaktfullApp(
-      title: 'Impaktfull UI Example',
-      home: HomeScreen(),
-    ));
+void main() => runApp(
+      const ImpaktfullApp(
+        title: 'Impaktfull UI Example',
+        home: HomeScreen(),
+      ),
+    );