+++ example/lib/src/screen/components/buttons_screen.dart
@@ -8,26 +8,28 @@ class ButtonsScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ImpaktfullScreen(
-      title: 'Components - Buttons',
-      onBackTapped: () => Navigator.of(context).pop(),
-      child: ImpaktfullListView(
-        children: [
-          ImpaktfullButton.primary(
-            label: 'Primary Button',
-            onTap: () => SnackyUtil.show('On Primary tapped'),
-          ),
-          const SizedBox(height: 8),
-          ImpaktfullButton.secondary(
-            label: 'Secondary Button',
-            onTap: () => SnackyUtil.show('On Secondary tapped'),
-          ),
-          const SizedBox(height: 8),
-          ImpaktfullButton.accent(
-            label: 'Accent Button',
-            onTap: () => SnackyUtil.show('On Accent tapped'),
-          ),
-        ],
+    return ImpaktfullThemeLocalizer(
+      builder: (context, teme) => ImpaktfullScreen(
+        title: 'Components - Buttons',
+        onBackTapped: () => Navigator.of(context).pop(),
+        child: ImpaktfullListView(
+          children: [
+            ImpaktfullButton.primary(
+              label: 'Primary Button',
+              onTap: () => SnackyUtil.show('On Primary tapped'),
+            ),
+            const SizedBox(height: 8),
+            ImpaktfullButton.secondary(
+              label: 'Secondary Button',
+              onTap: () => SnackyUtil.show('On Secondary tapped'),
+            ),
+            const SizedBox(height: 8),
+            ImpaktfullButton.accent(
+              label: 'Accent Button',
+              onTap: () => SnackyUtil.show('On Accent tapped'),
+            ),
+          ],
+        ),
       ),
     );
   }