+++ example/lib/src/screen/components/listview_screen.dart
@@ -10,29 +10,31 @@ class ListViewScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ImpaktfullScreen(
-      title: 'Components - ListView',
-      onBackTapped: () => Navigator.of(context).pop(),
-      child: ImpaktfullListView(
-        children: [
-          ImpaktfullButton.primary(
-            label: 'Children',
-            onTap: () => Navigator.of(context).push(MaterialPageRoute(
-                builder: (context) => const ListViewChildren())),
-          ),
-          const SizedBox(height: 8),
-          ImpaktfullButton.primary(
-            label: 'Item Builder',
-            onTap: () => Navigator.of(context).push(MaterialPageRoute(
-                builder: (context) => const ListViewItemBuilder())),
-          ),
-          const SizedBox(height: 8),
-          ImpaktfullButton.primary(
-            label: 'Item Separated Builder',
-            onTap: () => Navigator.of(context).push(MaterialPageRoute(
-                builder: (context) => const ListViewItemSeparatedBuilder())),
-          ),
-        ],
+    return ImpaktfullThemeLocalizer(
+      builder: (context, theme) => ImpaktfullScreen(
+        title: 'Components - ListView',
+        onBackTapped: () => Navigator.of(context).pop(),
+        child: ImpaktfullListView(
+          children: [
+            ImpaktfullButton.primary(
+              label: 'Children',
+              onTap: () => Navigator.of(context).push(MaterialPageRoute(
+                  builder: (context) => const ListViewChildren())),
+            ),
+            const SizedBox(height: 8),
+            ImpaktfullButton.primary(
+              label: 'Item Builder',
+              onTap: () => Navigator.of(context).push(MaterialPageRoute(
+                  builder: (context) => const ListViewItemBuilder())),
+            ),
+            const SizedBox(height: 8),
+            ImpaktfullButton.primary(
+              label: 'Item Separated Builder',
+              onTap: () => Navigator.of(context).push(MaterialPageRoute(
+                  builder: (context) => const ListViewItemSeparatedBuilder())),
+            ),
+          ],
+        ),
       ),
     );
   }