+++ example/lib/src/screen/home_screen.dart
@@ -14,8 +14,7 @@ class HomeScreen extends StatelessWidget {
         children: [
           ImpaktfullButton.primary(
             label: 'Components',
-            onTap: () => Navigator.of(context).push(MaterialPageRoute(
-                builder: (context) => const ComponentsScreen())),
+            onTap: () => Navigator.of(context).push(MaterialPageRoute(builder: (context) => const ComponentsScreen())),
           ),
         ],
       ),